backend/requirements.txt
fastapi
uvicorn
python-dotenv
snowflake-connector-python
pydantic
passlib[bcrypt]
jwt
